Unify Shortcut model
