+++ macOS/Accelerate/StoreManager.swift
